chore(app): Refactor app error logic
