+++ test/unit/machine-learning/machine-learning-api-client.spec.ts
@@ -23,7 +23,8 @@ import { FirebaseMachineLearningError } from '../../../src/machine-learning/mach
 import { HttpClient } from '../../../src/utils/api-request';
 import * as utils from '../utils';
 import * as mocks from '../../resources/mocks';
-import { FirebaseAppError, toHttpResponse } from '../../../src/utils/error';
+import { toHttpResponse } from '../../../src/utils/error';
+import { FirebaseAppError } from '../../../src/app/error';
 import { FirebaseApp } from '../../../src/app/firebase-app';
 import { getMetricsHeader, getSdkVersion } from '../../../src/utils/index';
 import { MachineLearningApiClient } from '../../../src/machine-learning/machine-learning-api-client';
